fix aws

+++ src/core/usecases/workspace/add-img/AddimgWorkspaceHandler.ts
@@ -7,43 +7,52 @@ import { CommandHandler } from "../../../shared/usecase/CommandHandler";
 import { AddimgWorkspaceInput } from "./AddimgWorkspaceInput";
 import { AddimgWorkspaceOutput } from "./AddimgWorkspaceOutput";
 import * as fs from "fs";
-
-
+import mime from "mime";
+import { SystemError } from "../../../shared/exceptions/SystemError";
+import { MessageError } from "../../../shared/exceptions/message/MessageError";
 
 @Service()
 export class AddimgWorkspaceHandler extends CommandHandler<
-AddimgWorkspaceInput,
-AddimgWorkspaceOutput
->{
-    constructor(
-        // @Inject("storage.service")
-        // private readonly _storageService: IStorageService,
+  AddimgWorkspaceInput,
+  AddimgWorkspaceOutput
+> {
+  constructor(
+    // @Inject("storage.service")
+    // private readonly _storageService: IStorageService,
 
-        @Inject("workspace.repository")
-        private readonly _workspaceRepository: IWorkSpaceRepository
+    @Inject("workspace.repository")
+    private readonly _workspaceRepository: IWorkSpaceRepository
+  ) {
+    super();
+  }
 
-    ){
-        super()
+  async handle(
+    id: string,
+    param: AddimgWorkspaceInput
+  ): Promise<AddimgWorkspaceOutput> {
+    await validateDataInput(param);
+    await validateDataInput(param);
+
+    const file = param.file;
+    const ext = mime.getExtension(file.mimetype);
+    if (!ext) {
+      throw new SystemError(MessageError.PARAM_INVALID, "image");
     }
 
-    async handle(id:string, param:  AddimgWorkspaceInput): Promise<AddimgWorkspaceOutput> {
-        await validateDataInput(param)
-        
-        const data = new WorkSpace()
-        const fileKey = WorkSpace.generateImg(param.file)
-        //  await this._storageService.uploadFile(param.file, fileKey);
-         fs.unlink(`uploads/${fileKey}`, (err)=>{
-           if(err) console.log(err);
-           
-         })
-        //  const readStream = await this._storageService.getFile(fileKey)
+    WorkSpace.validateImageFile(file);
+    const fileKey = WorkSpace.getImagePath(id, ext);
+    const data = new WorkSpace();
+    //  await this._storageService.uploadFile(param.file, fileKey);
+    fs.unlink(`uploads/${fileKey}`, (err) => {
+      if (err) console.log(err);
+    });
+    //  const readStream = await this._storageService.getFile(fileKey)
 
-         data.image = 'dsad';
+    data.image = fileKey;
 
-         const hasSucess = await this._workspaceRepository.update(id,data)
-         const result = new AddimgWorkspaceOutput()
-         result.setData(hasSucess)
-         return result
-    }
-
+    const hasSucess = await this._workspaceRepository.update(id, data);
+    const result = new AddimgWorkspaceOutput();
+    result.setData(hasSucess);
+    return result;
+  }
 }
